vista lista comidas

+++ demo/src/main/java/com/example/demo/entities/Comida.java
@@ -9,7 +9,7 @@ public class Comida {
     private Integer id;
     private String nombre;
     private String descripcion;
-    private Double precio;
+    private Integer precio;
     private String imagenUrl;
 }
